+++ docs/changelog.md
@@ -1,8 +1,8 @@
 # Changelog for `jupyterlab_latex`
 
-- **3.2.0**:
+- **4.1.0**:
 
-  - Added UI improvements to LaTeX Editor
+- Added UI improvements to LaTeX Editor
     - Preview Button
     - Symbol and Constant Menu
     - Filtered Error Messages
@@ -10,6 +10,12 @@
     - Equation Buttons
     - Typeface Modifier GUI
     - Bullet point and Numbered List Buttons
+
+- **3.2.0**:
+
+  - Minimum supported JupyterLab is 3.6.0
+  - Bump JS and Python dependencies
+  - Fix build issues
 
 * **3.1.0**: